Small bug fixes. Split HTTP export into two levels.
darcs-hash:20090429051525-20ca2-7025686beedd381e3b5f32c6ee86f3d4fb4ceb93.gz

+++ src/include/packet.h
@@ -111,7 +111,7 @@ struct struct_property_t {
   /** This represents the number of bytes from the begining of the
       struct where this item may be found 
   */
-  uint64_t item;
+  int item;
 
   int size;
